--- src/Api/src/WebScrap.API/Data/CssTagRanges.cs
@@ -1,5 +1,0 @@
-using System.Collections.Immutable;
-
-namespace WebScrap.API.Data;
-
-public record struct CssTagRanges(string Css, ImmutableArray<Range> TagRanges);
