--- src/Api/src/WebScrap.API/Data/IScrapResult.cs
@@ -1,8 +1,0 @@
-using System.Text.Json.Nodes;
-
-namespace WebScrap.API.Data;
-
-public interface IScrapResult
-{
-    JsonArray AsJson();
-}
