--- src/Api/src/WebScrap.API/Formatters/JsonFormatter.cs
@@ -1,23 +1,0 @@
-using System.Text.Json;
-using System.Text.Json.Nodes;
-
-namespace WebScrap.API.Formatters;
-
-internal static class JsonFormatter
-{
-    internal static JsonArray Format((string, JsonArray)[] arrays)
-    {
-        var result = new List<JsonNode>();
-        foreach (var item in arrays)
-        {
-            var obj = new
-            {
-                key = item.Item1,
-                values = item.Item2
-            };
-            result.Add(JsonSerializer.SerializeToNode(obj)!);
-        }
-
-        return new([..result]);
-    }
-}
