--- src/Api/src/WebScrap.API/WebScrap.API.csproj
@@ -1,33 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Core\Tags\WebScrap.Core.Tags\WebScrap.Core.Tags.csproj" />
-    <ProjectReference Include="..\..\..\Core\Css\src\WebScrap.Css.API\WebScrap.Css.API.csproj" />
-    <ProjectReference Include="..\..\..\Modules\WebScrap.Modules.Exporting.Json\WebScrap.Modules.Exporting.Json\WebScrap.Modules.Exporting.Json.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Formatters/" />
-  </ItemGroup>
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <LangVersion>preview</LangVersion>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageId>WebScrap.Json</PackageId>
-    <Version>1.0.0-rc.2</Version>
-    <Authors>Alex Kozachenko</Authors>
-    <PackageReadmeFile>Readme.md</PackageReadmeFile>
-    <PackageLicenseFile>License</PackageLicenseFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="..\..\..\..\Readme.md" Pack="true" PackagePath="\" />
-    <None Include="..\..\..\..\License" Pack="true" PackagePath="\" />
-  </ItemGroup>
-
-</Project>
