--- src/Api/src/src.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebScrap.API", "WebScrap.API\WebScrap.API.csproj", "{A3ED2EF9-CB20-4996-B01B-0772B939D656}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A3ED2EF9-CB20-4996-B01B-0772B939D656}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A3ED2EF9-CB20-4996-B01B-0772B939D656}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A3ED2EF9-CB20-4996-B01B-0772B939D656}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A3ED2EF9-CB20-4996-B01B-0772B939D656}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {024D2548-158E-47E6-A76B-5F541BFE0EC2}
-	EndGlobalSection
-EndGlobal
