--- src/Api/tests/WebScrap.API.KnownIssues.Tests/Categories.cs
@@ -1,6 +1,0 @@
-namespace WebScrap.API.KnownIssues.Tests;
-
-public static class Categories
-{
-    public const string KnownIssues = "KnownIssues";
-}
