--- src/Api/tests/WebScrap.API.Tests/Categories.cs
@@ -1,6 +1,0 @@
-namespace WebScrap.API.Tests;
-
-public static class Categories
-{
-    public const string Features = "Features";
-}
