--- src/Api/tests/WebScrap.API.Tests/GlobalUsings.cs
@@ -1,2 +1,0 @@
-global using NUnit.Framework;
-global using WebScrap.API;
