--- src/Api/tests/tests.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebScrap.API.KnownIssues.Tests", "WebScrap.API.KnownIssues.Tests\WebScrap.API.KnownIssues.Tests.csproj", "{4627721C-C19B-4705-9A01-3BA919110A34}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebScrap.API.Tests", "WebScrap.API.Tests\WebScrap.API.Tests.csproj", "{AB737E88-E859-4F94-ABA2-4EB076F44F6B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScrap.API", "..\src\WebScrap.API\WebScrap.API.csproj", "{DED93970-C610-437E-B376-286C83119528}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4627721C-C19B-4705-9A01-3BA919110A34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4627721C-C19B-4705-9A01-3BA919110A34}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4627721C-C19B-4705-9A01-3BA919110A34}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4627721C-C19B-4705-9A01-3BA919110A34}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AB737E88-E859-4F94-ABA2-4EB076F44F6B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AB737E88-E859-4F94-ABA2-4EB076F44F6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AB737E88-E859-4F94-ABA2-4EB076F44F6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AB737E88-E859-4F94-ABA2-4EB076F44F6B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DED93970-C610-437E-B376-286C83119528}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DED93970-C610-437E-B376-286C83119528}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DED93970-C610-437E-B376-286C83119528}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DED93970-C610-437E-B376-286C83119528}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A50AF384-B631-4963-9C98-16219B728B8E}
-	EndGlobalSection
-EndGlobal
