+++ src/WebScrap/src/WebScrap/Data/CssTagRanges.cs
@@ -1,0 +1,5 @@
+using System.Collections.Immutable;
+
+namespace DevOvercome.WebScrap.Data;
+
+public record struct CssTagRanges(string Css, ImmutableArray<Range> TagRanges);
