+++ src/WebScrap/src/WebScrap/Data/IScrapResult.cs
@@ -1,0 +1,8 @@
+using System.Text.Json.Nodes;
+
+namespace DevOvercome.WebScrap.Data;
+
+public interface IScrapResult
+{
+    JsonArray AsJson();
+}
