+++ src/WebScrap/src/WebScrap/WebScrap.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\Core\Tags\WebScrap.Core.Tags\WebScrap.Core.Tags.csproj" />
+    <ProjectReference Include="..\..\..\Core\Css\src\WebScrap.Css.API\WebScrap.Css.API.csproj" />
+    <ProjectReference Include="..\..\..\Modules\WebScrap.Modules.Exporting.Json\WebScrap.Modules.Exporting.Json\WebScrap.Modules.Exporting.Json.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Formatters/" />
+  </ItemGroup>
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <LangVersion>preview</LangVersion>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+</Project>
