+++ src/WebScrap/src/src.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScrap", "WebScrap\WebScrap.csproj", "{93B9BC72-F09E-4F49-932D-2230A2F4086C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{93B9BC72-F09E-4F49-932D-2230A2F4086C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93B9BC72-F09E-4F49-932D-2230A2F4086C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93B9BC72-F09E-4F49-932D-2230A2F4086C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93B9BC72-F09E-4F49-932D-2230A2F4086C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
