+++ src/WebScrap/tests/WebScrap.KnownIssues.Tests/Categories.cs
@@ -1,0 +1,6 @@
+namespace WebScrap.KnownIssues.Tests;
+
+public static class Categories
+{
+    public const string KnownIssues = "KnownIssues";
+}
