+++ src/WebScrap/tests/WebScrap.Tests/Categories.cs
@@ -1,0 +1,6 @@
+namespace WebScrap.Tests;
+
+public static class Categories
+{
+    public const string Features = "Features";
+}
