+++ src/WebScrap/tests/WebScrap.Tests/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using NUnit.Framework;
+global using DevOvercome.WebScrap;
